schema_migration/python/requirements.txt
fastavro
requests
kafka-python
pyyaml